// frontend/src/lib/api/orval/api/generated/model/mCQExercise.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Langu-AI API
 * API for the Langu-AI German language learning platform
 * OpenAPI spec version: 1.0.0
 */
import type { MCQExerciseExerciseType } from "./mCQExerciseExerciseType";
import type { MCQChoice } from "./mCQChoice";

export interface MCQExercise {
  readonly id: string;
  exercise_type: MCQExerciseExerciseType;
  readonly item_id: string;
  readonly item_text: string;
  readonly item_translation: string;
  question: string;
  explanation?: string;
  readonly choices: readonly MCQChoice[];
  readonly correct_choice_id: string;
}